Import project to linux

+++ src/XiEngine/math/vector3.cpp
@@ -115,7 +115,7 @@ float Vector3::lengthSquared() const
 
 Vector3 Vector3::abs() const
 {
-	return Vector3(std::fabsf(x_), std::fabsf(y_), std::fabsf(z_));
+	return Vector3(std::fabs(x_), std::fabs(y_), std::fabs(z_));
 }
 	
 Vector3& Vector3::normalize()
